requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
cloudpickle==3.1.1
dask==2025.2.0
distro==1.9.0
evtx==0.8.8
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
llm==0.21
locket==1.0.0
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.63.0
packaging==24.2
pandas==2.2.3
partd==1.4.2
pluggy==1.5.0
puremagic==1.28
pydantic==2.10.6
pydantic_core==2.27.2
pyfiglet==1.0.2
pyparsing==3.2.1
pySigma==0.11.19
pysigma-backend-carbonblack==0.1.8
pysigma-backend-cortexxdr==0.1.4
pySigma-backend-kusto==0.4.3
pytest==8.3.4
python-dateutil==2.9.0.post0
python-ulid==3.0.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
tabulate==0.9.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
